+++ NEWS.md
@@ -1,7 +1,11 @@
-# psichomics 1.12.1 (17 January, 2020)
+# psichomics 1.12.1 (21 January, 2020)
 
-* Fix loading only a portion of TCGA data (with no warning) if too many cohorts
-are selected
+* Loading of TCGA and SRA data:
+    - All supported TCGA data can now be completely loaded when all cohorts are
+    selected instead of loading only a portion of TCGA data without any warning
+    - SRA projects containing only one column of extra information in sample
+    metadata are now correctly loaded instead of raising an error
+    (`loadSRAproject()`)
 * User-provided junction quantification loading:
     - Support junction coordinates from mitochondrial, Z and W chromosomes
     - Fix issues with files containing splice junctions within random, alt and